+++ spring-kafka/src/main/java/org/springframework/kafka/requestreply/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Provides classes for request/reply semantics.
  */
-package org.springframework.kafka.sendreceive;
+package org.springframework.kafka.requestreply;
